refactor resource handling related operations out of mappings.py
