+++ bagel/mappings.py
@@ -56,43 +56,6 @@ CONFIG_NAMESPACES_PATH = (
 CONFIG_NAMESPACES_URL = "https://raw.githubusercontent.com/neurobagel/communities/refs/heads/main/config_metadata/config_namespace_map.json"
 
 
-def get_available_configs(config_namespaces_mapping: list) -> list:
-    """Return the list of names of available community configurations."""
-    return [config["config_name"] for config in config_namespaces_mapping]
-
-
-def get_supported_namespaces_for_config(config_name: str) -> dict:
-    """Return a dictionary of supported namespace prefixes and their corresponding full URLs for a given community configuration."""
-    config_namespaces = next(
-        config["namespaces"]
-        for config in CONFIG_NAMESPACES_MAPPING
-        if config["config_name"] == config_name
-    )
-
-    config_namespaces_dict = {}
-    for namespace_group in config_namespaces.values():
-        for namespace in namespace_group:
-            config_namespaces_dict[namespace["namespace_prefix"]] = namespace[
-                "namespace_url"
-            ]
-
-    return config_namespaces_dict
-
-
-def parse_pipeline_catalog(pipeline_catalog: list) -> tuple[dict, dict]:
-    """
-    Load the pipeline catalog and return a dictionary of pipeline names and their URIs in the Nipoppy namespace,
-    and a dictionary of pipeline names and their supported versions in Nipoppy.
-    """
-    version_dict = {}
-    uri_dict = {}
-    for pipeline in pipeline_catalog:
-        version_dict[pipeline["name"]] = pipeline["versions"]
-        uri_dict[pipeline["name"]] = f"{NP.pf}:{pipeline['name']}"
-
-    return uri_dict, version_dict
-
-
 # TODO: consider refactoring this into a Mappings class that also
 # handles lazy loading of the remote content, i.e. only when accessed
 PIPELINE_CATALOG, PIPELINES_FETCHING_ERR = file_utils.request_file(
