+++ tests/unit/test_pheno_utils.py
@@ -6,6 +6,54 @@ from bagel import mappings
 from bagel.utilities import pheno_utils
 
 
+@pytest.fixture(scope="session")
+def mock_config_namespaces_mapping():
+    return [
+        {
+            "config_name": "Neurobagel",
+            "namespaces": {
+                "variables": [
+                    {
+                        "namespace_prefix": "nb",
+                        "namespace_url": "http://neurobagel.org/vocab/",
+                    }
+                ],
+                "terms": [
+                    {
+                        "namespace_prefix": "ncit",
+                        "namespace_url": "http://ncicb.nci.nih.gov/xml/owl/EVS/Thesaurus.owl#",
+                    },
+                    {
+                        "namespace_prefix": "snomed",
+                        "namespace_url": "http://purl.bioontology.org/ontology/SNOMEDCT/",
+                    },
+                ],
+            },
+        },
+        {
+            "config_name": "Ontario Brain Institute",
+            "namespaces": {
+                "variables": [
+                    {
+                        "namespace_prefix": "nb",
+                        "namespace_url": "http://purl.bioontology.org/ontology/MEDDRA/",
+                    }
+                ],
+                "terms": [
+                    {
+                        "namespace_prefix": "lnc",
+                        "namespace_url": "http://purl.bioontology.org/ontology/LNC/",
+                    },
+                    {
+                        "namespace_prefix": "medra",
+                        "namespace_url": "http://purl.bioontology.org/ontology/MEDDRA/",
+                    },
+                ],
+            },
+        },
+    ]
+
+
 @pytest.mark.parametrize(
     "partial_data_dict, invalid_column_name",
     [
@@ -698,3 +746,27 @@ def test_additional_config_help_text(monkeypatch):
         "Failed to locate any community configurations."
         in pheno_utils.additional_config_help_text()
     )
+
+
+def test_get_available_configs(mock_config_namespaces_mapping):
+    """Test the function returns a correct list of config names from a configuration namespaces mapping file."""
+    assert pheno_utils.get_available_configs(
+        mock_config_namespaces_mapping
+    ) == [
+        "Neurobagel",
+        "Ontario Brain Institute",
+    ]
+
+
+def test_get_supported_namespaces_for_config(
+    mock_config_namespaces_mapping, monkeypatch
+):
+    """Test the function correctly returns the supported namespaces for a given config name."""
+    monkeypatch.setattr(
+        mappings, "CONFIG_NAMESPACES_MAPPING", mock_config_namespaces_mapping
+    )
+    assert pheno_utils.get_supported_namespaces_for_config("Neurobagel") == {
+        "nb": "http://neurobagel.org/vocab/",
+        "ncit": "http://ncicb.nci.nih.gov/xml/owl/EVS/Thesaurus.owl#",
+        "snomed": "http://purl.bioontology.org/ontology/SNOMEDCT/",
+    }
